Reintroduce attribute decoding by tags
Previously the function pointer was encoded directly into the attribute
map. The impetus for this was a hope for improved decoding times, but
performance didn't tangibly change and I kept the function pointer
methods. I now realize that attribute tags bring much needed
insight when debugging as the `Debug` trait can be trivially derived for
attribute tags, but not for function pointers. Thus, with no performance
hit and a helping hand with debugging, attribute tags have been
reintroduced.