Update crc to slice by 16
This presents a 2x throughput increase compared to slice by 8
originally.

This commit also moves the crc table generation to build time to save
allow for easier migration to a different slice by length in the future.
It also cuts down on the amount of committed code. This had a neglible
effect on compile times