Support multi-id objects
Every replay contains a list of objects like
"TAGame.GameEvent_Soccar_TA". The index of the object in the list is
used throughout the replay as the object's id. This works in the vast
majority of replays, but there are some replays where an object is
listed multiple times in the list. Then when resolving the available
attributes for an object, only the first occurences object id was used,
resulting in replay errors.

The fix is to consider every object id for a given name like
"TAGame.GameEvent_Soccar_TA" when resolving attributes.