Decode rigid body rotation bits into quaternions
While boxcars could parse all replays, it nonsensically stored the quaternion
information from rigid body attributes. This commit digests the logic
from rattletrap, bakkes, and jjbott's implementation. Output of the
ported logic has been tested against bakkes to ensure correctness.

Benchmarks show a small, but non-negligible performance decrease (1-2%)
due to the increase in number of calculations when decoding a rigid body
(the most common attribute).

This is a breaking change as the RigidBody attribute changed it's
structure.

Thank you to the carball devs for identifying the issue.