Replace vector bias with computed values
Previously boxcars would output the following vector data for rigid
bodies:

```
"location": {
  "bias": 262144,
  "dx": 457343,
  "dy": 15746,
  "dz": 263845
}
```

This is correct, but does not conform to the other parsers (bakkes and
jjbott) which look like:

```
"position": {
  "X": 1951.99,
  "Y": -2463.98,
  "Z": 17.01
},
```

The vector fields are computed as ((X - bias) / 100). Interestingly
enough, the `/ 100` part of the formula is dropped for new actors (I do
not know why). Rattletrap also never computes `/ 100`. I'm not
completely sure this is the correct approach, but it's better to conform
here.

Having floating point and integer vectors necessitated splitting the
Vector class into two: Vector3f and Vector3i (one for integer and the
other for floating point -- side note: Vectorf and Vectori just didn't
look right).

This is a breaking change to a lot of attributes and new actors. So I'll
be creating a release candidate before cutting an official (minor)
release.

Another tidbit: despite the additional calculation, this commit improves
performance about 2%. Probably due to less data needing to be allocated,
as the Vector structs decreased 4 bytes in size.