Move large network attributes to heap
The Attribute enum is large. A single instance, consumes 168 bytes. When
most of the attributes are rigid bodies (64 bytes) this consumes a
significant amount of empty space (168 - 64). By moving all attributes
that are larger than a rigid body to the heap, throughput benchmarks
improved 40%.

This technically a breaking change, due to the attribute type changing
(from inline to `Box`), but since boxes can be easily dereferenced, this
should be a non-issue.