Migrate StatEvent to struct with signed object id field
The stat event attribute used to be a tuple with an unsigned field. The
tuple isn't descriptive of the available fields and the unsigned field
would display (2^32 - 1), which is incorrect. Therefore this commit
migrates to a struct with named fields with the object id as a signed
integer. I would have liked to have used `ObjectId`, but that is an
unsigned value so it is not as easy to slot in.