Stable and efficient header prop serialization
I was looking to do snapshot testing of replay JSON output to ensure
stability for any future changes, but I realized that the output
slightly changed from run to run.

The reason why is due to the serialization of header properties that
used an intermediate hashmap, which doesn't guarantee iteration order
from one run to the next.

By getting rid of the hashmap we improve efficiency and guarantee stable
output.