Add snapshot testing
By creating a snapshot of a hash of the json serialized data, it will be
obvious what the ramifications are for changing any bit of code.

This should allow a bit more confidence when making code changes.

The commit also adds in a json command to dump a replay json to stdout