Remove header property special casing with quirks mode
See #165 for more details
Closes #165

The short of it, when a version of `0.0.None` is encountered, the header
property decoder enters quirks mode, which eliminates some hard coded
checks.

No change in output (thank you snapshot tests). No change in
performance.

Since there was no change in output, I'm ok removing a test as I'm too
lazy to retool it.