Branchless buffer refills for vector decoding
This reduces the number instructions emitted by nearly half when we
check if there is enough data to blindly refill the lookahead. The best
part, we don't even need to use the unchecked API to blindly refill --
somehow the compiler can figured this out when we ensure there is at
least 16 bytes left unbuffered. Crazy.