+25% throughput with segmented arrays (#188)
* +25% throughput with segmented arrays

For every updated actor in the network body, there are two hash
functions run. One to lookup the actor id to grab their object id, and
another to look up that object id's attributes. Even with the fastest
hash function (FNV), this is still slow when executed a 100k times in a
replay.

The solution is to transition the hashmap to a segmented array where
small indices (which are the vast majority of them) are looked up
directly in a sparse array with a static size. Indices that fall outside
the array are still stored in a hashmap.

Benchmarks showed a +25% throughput improvement, which imo is kinda
incredible that such performance improvement are still on the table.

* ...

* update perf numbers

* Update error model