Reinstate vote actor support
In #185 I deleted code that fell back on the class hierarchy described
in the replay instead of the ones we hardcode. Turns out, when running a
regression test against the main replay corpus of 200k replays files --
2 replays depended only successfully parsed if this fallback was
present. I've added the hard coded the missing vote actor, but the main
thing is the adding the regression replay to the codebase so this
doesn't happen again in the future.

https://github.com/nickbabcock/boxcars/commit/0ce77fda081404f7ca41ddb9ad548aa1c500637e#diff-bc565c9df4b377f07d29b6132b181f1a239dfd8acab3700841961ed4550f8975L139-L150