Ignore size when parsing header properties
In #165, it was discussed that when decoding a property there is a size
prefix. So in #167, I introduced the concept of a sub-parser so to
ensure that this size limit isn't overrun.

However this is not always the case, and UTF16 strings appear to be able
to overrun this size, so a test case has been added to prevent further
regressions. Ref #190